List requests and do action

+++ app/Models/UserProfile.php
@@ -43,12 +43,12 @@ class UserProfile extends Model
         return $this;
     }
 
-    public function getRating(): int
+    public function getRating(): float
     {
         return $this->rating;
     }
 
-    public function setRating(int $rating): self
+    public function setRating(float $rating): self
     {
         $this->rating = $rating;
         return $this;
