+++ database/migrations/2023_03_23_190312_create_user_profiles_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('user_profiles', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users');
-            $table->tinyInteger('rating')->default(0);
+            $table->decimal('rating', 2, 1)->default(0);
             $table->timestamps();
         });
     }
